Update Z

Update z_vec in the ADMM algorithm

RSAVS_UpdateZ_L1(invec, param, r1, const_a)

RSAVS_UpdateZ_L2(invec, param, r1, const_a)

RSAVS_UpdateZ_Huber(invec, param, r1, const_a)

Arguments

invec
numerical vector, the orignal z_vec

param
numerical vector, parameters needed for the loss function. For "L1" and "L2" loss, this will be ignored

r1
numerical scalar, parameter needed in the quadratic term in the augmented part for z_vec.

const_a
numerical scalar, parameter controls the weight of regression part in the orignal objective function

Details

These functions correspond to different types of loss function defined in the original problem.

RSAVS_UpdateZ_L1 corresponds to the "L1" loss, which means l_type = "L1".

RSAVS_UpdateZ_L2 corresponds to the "L2" loss, which means l_type = "L2".

RSAVS_UpdateZ_Huber corresponds to the "Huber" loss, which means l_type = "Huber".

Note

When the loss type is "L2", it is possible to modify the algorithm to proceed without z_vec. But currently for a more unified paradigm, this is not implemented.

Please refer to the vignette about the algorithm detail design to find out more about how these parameters are defined.

See also