Built-in penalty functions

These are built-in penalty functions. They are implemented purely in R.

penalty_lasso(x, params, derivative = FALSE)

penalty_scad(x, params, derivative = FALSE)

penalty_mcp(x, params, derivative = FALSE)

Arguments

x
input numeric vector

params
numerical vector, parameters needed for the corresponding penalty function. Refer to the following sections for more details.

derivative
boolen, whether the derivatives at x should be computed instead of the actual penalized value. Default to FALSE.

Value

A numerical vector storing the penalty function's value at x if derivative = FALSE. Otherwise the derivative of the penalty function at x

Parameters for different penalties

For lasso, $$ f(x, \lambda) = \lambda |x| $$ where \(\lambda\) = param[1].

For SCAD and MCP, \(\lambda\) = param[1] and \(\gamma\) = param[2]

Examples

x <- seq(from = -10, to = 10, by = 0.001)
lasso <- penalty_lasso(x, 1.0)

#> Error in penalty_lasso(x, 1): could not find function "penalty_lasso"

dlasso <- penalty_lasso(x, 1.0, TRUE)

#> Error in penalty_lasso(x, 1, TRUE): could not find function "penalty_lasso"

scad <- penalty_scad(x, c(1.5, 3.7))

#> Error in penalty_scad(x, c(1.5, 3.7)): could not find function "penalty_scad"

dscad <- penalty_scad(x, c(1.5, 3.7), TRUE)

#> Error in penalty_scad(x, c(1.5, 3.7), TRUE): could not find function "penalty_scad"

mcp <- penalty_mcp(x, c(1.5, 2.7))

#> Error in penalty_mcp(x, c(1.5, 2.7)): could not find function "penalty_mcp"

dmcp <- penalty_mcp(x, c(1.5, 2.7), TRUE)

#> Error in penalty_mcp(x, c(1.5, 2.7), TRUE): could not find function "penalty_mcp"

  
matplot(x, cbind(lasso, scad, mcp), type = "l", ylab = "penalty")

#> Error in cbind(lasso, scad, mcp): object 'lasso' not found

legend(-10, 2.5, legend = c("lasso", "scad", "mcp"), col = 1 : 3, lty = 1 : 3)

#> Error in strwidth(legend, units = "user", cex = cex, font = text.font): plot.new has not been called yet

matplot(x, cbind(dlasso, dscad, dmcp), type = "l", ylab = "derivative")

#> Error in cbind(dlasso, dscad, dmcp): object 'dlasso' not found

legend(-10, 1.5, legend = c("lasso", "scad", "mcp"), col = 1 : 3, lty = 1 : 3)

#> Error in strwidth(legend, units = "user", cex = cex, font = text.font): plot.new has not been called yet