Post-selection estimation

This function carries out the post-selection estimation given the model structure.

RSAVS_Further_Improve(
  y_vec,
  x_mat,
  l_type = "1",
  l_param = NULL,
  mu_vec,
  beta_vec
)

Arguments

y_vec
numeric vector for response. n = length(y_vec) is the number of observations.

x_mat
numeric matrix for covariates. Each row is for one observation. p = ncol(x_mat) is the number of covariates.

l_type
character, type of loss function.

"1": L-1 loss.

"2": L-2 loss.

"H": Huber loss.

The default value is "1".

l_param
vector of parameters needed by the loss function. For Huber loss, c = l_param[1] and a popular choice is c = 1.345. The Default value is NULL since no additional parameters are needed for the default L-1 loss.

mu_vec
a length-n vector for subgroup effect. The function uses this to determine the subgroup structure.

beta_vec
a length-p vector for covariate effect. The function uses this to find the active covariates.

Value

a list containing the improved mu(mu_vec) and beta(beta_vec) vector.

Details

This function uses mu_vec to determine the subgroup structure matrix via RSAVS_Mu_to_Mat. The active covariates are those with non-zero beta_vec entries.

Since this is the post-selection estimation. It's a regular estimate without any penalties. One should make sure the given model structure is identifiable.

See also