Robust subgroup analysis and variable selection simultaneously

This function utilize the cpp solver when carring out robust subgroup analysis and variable selection simultaneously. And it is the core solver for RSAVS_Path.

RSAVS_Solver(
  y_vec,
  x_mat,
  l_type = "L1",
  l_param = NULL,
  p1_type = "S",
  p1_param = c(2, 3.7),
  p2_type = "S",
  p2_param = c(2, 3.7),
  const_r123,
  const_abc = rep(1, 3),
  initial_values,
  additional,
  tol = 0.001,
  max_iter = 10,
  cd_max_iter = 1,
  cd_tol = 0.001,
  phi = 1,
  subgroup_benchmark = FALSE
)

Arguments

y_vec
numerical vector of response. n = length(y_vec) is the number of observations.

x_mat
numerical matrix of covariates. Each row for one observation and p = ncol(x_mat) is the number of covariates.

l_type
character string, type of loss function.

"L1": l-1 loss(absolute value loss)

"L2": l-2 loss(squared error loss)

"Huber": Huber loss. Its parameter is given in l_param.

Default value is "L1".

l_param
numeric vector containing necessary parameters of the corresponding loss function. The default value is NULL.

p1_type, p2_type
a character indicating the penalty types for subgroup identification and variable selection.

"S": SCAD

"M": MCP

"L": Lasso

Default values for both parameters are "S".

p1_param, p2_param
numerical vectors providing necessary parameters for the corresponding penalties.

For Lasso, lam = p_param[1]

For SCAD and MCP, lam = p_param[1], gamma = p_param[2]

Default values for both parameters are c(2, 3.7). Note: This function searches the whole lam1_vec * lam2_vec grid for the best solution. Hence the lambdas provided in these parameters serve only as placeholder and will be ignored and overwritten in the actual computation.

const_r123
a length-3 numerical vector, providing the scalars needed in the augmented lagrangian part of the ADMM algorithm

const_abc
a length-3 numeric vector, providing the scalars to adjust weight of regression function, penalty for subgroup identification and penalty for variable selection in the overall objective function. Defaults to c(1, 1, 1).

initial_values
list of vector, providing initial values for the algorithm.

additional
a list providing additional variables needed during the algorithm.

tol
numerical, convergence tolerance for the algorithm.

max_iter
integer, max number of iteration during the algorithm.

cd_max_iter
integer, max number of iteration during the coordinate descent update of mu and beta. If set to 0, will use analytical solution( instead of coordinate descent algorithm) to update mu and beta.

cd_tol
numerical, convergence tolerance for the coordinate descent part when updating mu and beta.

phi
numerical variable. A parameter needed for mBIC.

subgroup_benchmark
bool. Whether this call should be taken as a benchmark of subgroup identification. If TRUE, then the penalty for variable selection will be surpressed to a minimal value.

lam1_vec, lam2_vec
numerical vectors of customized lambda vectors. For lam1_vec, it's preferred to be in the order from small to big.

min_lam_ratio
the ratio between the minimal and maximal lambda, equals to (minimal lambda) / (maximal lambda). The default value is 0.03.

lam1_len, lam2_len
integers, lengths of the auto-generated lambda vectors.

See also