Update S and W

These functions perform the updating step of s_vec and w_vec in the ADMM algorithm for different types of penalty functions.

The updating steps of s_vec and w_vec share the same pattern in the ADMM algorithm. These functions correspond to different penalty types (p1_type and p2_type) provided from the original solver function.

UpdateSW_Identity: identity update, this is needed when there is no shrinkage applied, possibly since the corresponding lam is set to 0.

UpdateSW_Lasso: Lasso update

UpdateSW_SCAD: SCAD update

UpdateSW_MCP: MCP update

RSAVS_UpdateSW_Identity(invec, param, const_r, const_bc)

RSAVS_UpdateSW_Lasso(invec, param, const_r, const_bc)

RSAVS_UpdateSW_SCAD(invec, param, const_r, const_bc)

RSAVS_UpdateSW_MCP(invec, param, const_r, const_bc)

Arguments

invec

numerical vector, the orignal s_vec or w_vec.

param

numerical vector, parameters needed for the penalty function.

const_r

numerical scalar, parameter needed in the quadratic term in the augmented part.

const_bc

numerical scalar, parameter represents the weight of corresponding penalty part in the objective function

Note

Please refer to the vignette about the algorithm detail design to find out more about how these parameters are defined.

See also